Changed login label styling
