+++ Kick-Starter/TeamInformationWatch WatchKit Extension/InterfaceController.swift
@@ -47,21 +47,3 @@ extension InterfaceController: WCSessionDelegate {
         }
     }
 }
-
-extension WKInterfaceImage {
-    func loadImage(urlString: String) {
-        guard let url = URL(string: urlString) else {
-            return
-        }
-        self.setImage(UIImage(named: "logo"))
-        DispatchQueue.global().async { [weak self] in
-            if let data = try? Data(contentsOf: url) {
-                if let image = UIImage(data: data) {
-                    DispatchQueue.main.async {
-                        self?.setImage(image)
-                    }
-                }
-            }
-        }
-    }
-}
